requirements.txt
# Core dependencies for Streamlit dashboard
streamlit>=1.50.0
pandas>=2.2.2
PyMySQL>=1.1.1
altair>=5.0.0
python-dotenv>=1.0.1

